Model details and deployement/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
transformers
sentencepiece
git-lfs
python-multipart
